Hook up new instructions
